fix: refactor Tooltip to a controlled component
