+++ src/components/Tooltip/helpers/__test__/resolverCenter.spec.js
