+++ src/components/Tooltip/helpers/__test__/tooltipPositionResolver.spec.js
