+++ src/components/Tooltip/helpers/resolverCenter.js
