+++ src/components/Tooltip/index.d.ts
