+++ src/components/Tooltip/readme.md
@@ -1,26 +1,46 @@
 ##### Tooltip:
 
 ```js
-/* eslint-disable import/extensions */
-import React, { useRef } from 'react';
-import { Button } from 'react-rainbow-components';
-import showTooltip from '../../libs/showTooltip';
+/* eslint-disable import/extensions, import/no-unresolved */
+import React, { useRef, useState } from 'react';
+import { Button, Tooltip } from 'react-rainbow-components';
+import { useOutsideClick } from '../../libs/hooks';
+import { WindowScrolling } from '../../libs/scrollController';
 
 function Example() {
     const ref = useRef();
+    const tooltipRef = useRef();
+    const [isOpen, setIsOpen] = useState(false);
+
+    useOutsideClick(
+        tooltipRef,
+        () => {
+            setIsOpen(false);
+        },
+        isOpen,
+    );
+
+    const windowScrolling = new WindowScrolling();
+    windowScrolling.startListening(() => {
+        windowScrolling.stopListening();
+        setIsOpen(false);
+    });
+
     const show = () => {
-        showTooltip({ triggerElementRef: ref.current.htmlElementRef, content: (
-            <p>
-                123fsaddasd123fsaddasd123f
-                <div>saddasd123fsaddasd123fsaddasd123fsaddas d123fsaddasd123fsaddasd123fsaddasd123fsaddasd1</div>
-                <div>saddasd123fsaddasd123fsaddasd123fsaddas d123fsaddasd123fsaddasd123fsaddasd123fsaddasd1</div>
-            </p>
-        ) })
+        // setIsOpen(true);
+        setIsOpen(true);
     }
 
     return (
         <div>
             <Button style={{ marginBottom: 800, marginTop: 800 }} label="asd" onClick={show} ref={ref} />
+            <Tooltip triggerElementRef={() => ref.current.htmlElementRef} isVisible={isOpen} ref={tooltipRef}>
+                <p>
+                    123fsaddasd123fsaddasd123f
+                    <div>saddasd123fsaddasd123fsaddasd123fsaddas d123fsaddasd123fsaddasd123fsaddasd123fsaddasd1</div>
+                    <div>saddasd123fsaddasd123fsaddasd123fsaddas d123fsaddasd123fsaddasd123fsaddasd123fsaddasd1</div>
+                </p>
+            </Tooltip>
         </div>
     )
 }
